+++ src/components/Wrapper/Wrapper.tsx
@@ -1,29 +1,22 @@
 import React from 'react'
-import { composeClasses } from 'lib/classes'
+import { composeClasses } from '../../lib/classes'
+import { MaxWidth, Padding } from '../../interfaces/types'
+import { spacing } from '../../lib/spacing'
 
 export interface WrapperProps extends React.HTMLProps<HTMLDivElement> {
     children: React.ReactNode
     className?: string
-    paddingVertical?: number
-    paddingHorizontal?: number
-    maxWidth?: 'full' | 'screen-2xl' | 'screen-xl' | 'screen-lg' | 'screen-md' | 'screen-sm'
+    paddingY?: Padding
+    paddingX?: Padding
+    maxWidth?: MaxWidth
     hasViewportHeight?: boolean
 }
 
-function Wrapper({
-    children,
-    className,
-    paddingVertical = 10,
-    paddingHorizontal = 20,
-    maxWidth = 'screen-2xl',
-    hasViewportHeight,
-    style,
-    ...otherProps
-}: WrapperProps) {
+function Wrapper({ children, className, paddingY = '10', paddingX = '20', maxWidth = 'screen-2xl', hasViewportHeight, style, ...otherProps }: WrapperProps) {
     return (
         <div
             style={{ ...style, minHeight: hasViewportHeight ? 'calc(100vh - 193px)' : '' }}
-            className={composeClasses(`py-${paddingVertical} max-w-${maxWidth} px-${paddingHorizontal} mx-auto`, className)}
+            className={composeClasses(`max-w-${maxWidth}`, `py-${paddingY}`, `px-${paddingX}`, spacing.auto.marginHorizontal, className)}
             {...otherProps}
         >
             {children}