+++ src/lib/font.ts
@@ -1,0 +1,18 @@
+export const fontSize = {
+    xs: 'text-xs',
+    sm: 'text-sm',
+    md: 'text-base',
+    lg: 'text-lg',
+    xl: 'text-xl',
+    xxl: 'text-2xl',
+    xxxl: 'text-3xl',
+    jmb: 'text-4xl'
+}
+
+export const fontWeight = {
+    light: 'font-light',
+    normal: 'font-normal',
+    medium: 'font-medium',
+    semibold: 'font-semibold',
+    bold: 'font-bold'
+}