+++ src/lib/shape.ts
@@ -1,0 +1,37 @@
+export const borderRadius = {
+    none: {
+        left: 'rounded-l-none',
+        top: 'rounded-t-none',
+        right: 'rounded-r-none',
+        bottom: 'rounded-b-none',
+        all: 'rounded-none'
+    },
+    sm: {
+        left: 'rounded-l',
+        top: 'rounded-t',
+        right: 'rounded-r',
+        bottom: 'rounded-b',
+        all: 'rounded'
+    },
+    md: {
+        left: 'rounded-l-md',
+        top: 'rounded-t-md',
+        right: 'rounded-r-md',
+        bottom: 'rounded-b-md',
+        all: 'rounded-md'
+    },
+    lg: {
+        left: 'rounded-l-lg',
+        top: 'rounded-t-lg',
+        right: 'rounded-r-lg',
+        bottom: 'rounded-b-lg',
+        all: 'rounded-lg'
+    },
+    full: {
+        left: 'rounded-l-full',
+        top: 'rounded-t-full',
+        right: 'rounded-r-full',
+        bottom: 'rounded-b-full',
+        all: 'rounded-full'
+    }
+}